feat(service-sg-server): fix response for new component
https://github.com/akky-xxxx/project-my-homepage/issues/73
https://github.com/akky-xxxx/project-my-homepage/issues/101
